+++ packages/atproto-browser/app/at/route.ts
@@ -1,10 +1,14 @@
-import { getAtUriPath } from "@/lib/util";
-import { AtUri } from "@atproto/syntax";
-import { redirect } from "next/navigation";
+import { navigateAtUri } from "@/lib/navigation";
 
-export function GET(request: Request) {
+export async function GET(request: Request) {
   const searchParams = new URL(request.url).searchParams;
-  console.log(searchParams.get("u"));
-  const uri = new AtUri(searchParams.get("u")!);
-  redirect(getAtUriPath(uri));
+  const u = searchParams.get("u");
+  if (!u) {
+    return new Response("Missing u parameter", { status: 400 });
+  }
+  const result = await navigateAtUri(u);
+  if ("error" in result) {
+    return new Response(result.error, { status: 400 });
+  }
+  throw new Error("Should have redirected");
 }
